UniShed/requirements.txt
Django==3.0.5
django-phonenumber-field==4.0.0
phonenumbers==8.12.4
beautifulsoup4==4.9.1
django-bootstrap4==1.1.1
soupsieve==2.0.1
django-bootstrap-datepicker-plus==3.0.5
prettytable==0.7.2
django-multiselectfield==0.1.12
django-tables2==2.3.1
